requirements.txt
tensorflow-gpu>=1.12.0
matplotlib>=2.1.2
pandas>=0.22.0
pyqt5>=5.11.3
networkx>=2.2